rename dict.js to const.js
