+++ example/diam/const.js
